feat: add watched account as onboarding account creation option
