+++ apps/extension/src/ui/apps/onboard/routes/index.tsx
@@ -6,6 +6,7 @@ import { AccountAddDerivedPage } from "./AddAccount/AccountAddDerivedPage"
 import { AccountAddJsonPage } from "./AddAccount/AccountAddJsonPage"
 import { AccountAddLedgerOnboardWizard } from "./AddAccount/AccountAddLedgerWizard"
 import { AccountAddSecretOnboardWizard } from "./AddAccount/AccountAddSecretWizard"
+import { AccountAddWatchedPage } from "./AddAccount/AccountAddWatchedPage"
 import { PasswordPage } from "./Password"
 import { PrivacyPage } from "./Privacy"
 import { SuccessPage } from "./Success"
@@ -29,8 +30,8 @@ const OnboardingRoutes = () => {
             <Route path="json" element={<AccountAddJsonPage />} />
             <Route path="secret/*" element={<AccountAddSecretOnboardWizard />} />
             <Route path="ledger/*" element={<AccountAddLedgerOnboardWizard />} />
-            {/* <Route path="qr/*" element={<AccountAddQrWizard />} />
-            <Route path="watched" element={<AccountAddWatchedPage />} /> */}
+            {/* <Route path="qr/*" element={<AccountAddQrWizard />} /> */}
+            <Route path="watched" element={<AccountAddWatchedPage />} />
             <Route path="*" element={<Navigate to="" replace />} />
           </Route>
         </Route>
