+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddJson/index.tsx
@@ -1,11 +1,12 @@
 import { Spacer } from "@talisman/components/Spacer"
 
+import { AccountAddPageProps } from "../types"
 import { JsonAccountImportProvider } from "./context"
 import { ImportJsonAccountsForm } from "./ImportJsonAccountsForm"
 import { ImportJsonFileDrop } from "./ImportJsonFileDrop"
 import { UnlockJsonFileForm } from "./UnlockJsonFileForm"
 
-export const AccountAddJson = ({ onSuccess }: { onSuccess: (address: string) => void }) => {
+export const AccountAddJson = ({ onSuccess }: AccountAddPageProps) => {
   return (
     <JsonAccountImportProvider>
       <ImportJsonFileDrop />
