+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddLedger/index.tsx
@@ -1,10 +1,11 @@
 import { Navigate, Route, Routes } from "react-router-dom"
 
+import { AccountAddPageProps } from "../types"
 import { AddLedgerSelectAccount } from "./AddLedgerSelectAccount"
 import { AddLedgerSelectNetwork } from "./AddLedgerSelectNetwork"
 import { AddLedgerAccountProvider } from "./context"
 
-export const AccountAddLedgerWizard = ({ onSuccess }: { onSuccess: (address: string) => void }) => (
+export const AccountAddLedgerWizard = ({ onSuccess }: AccountAddPageProps) => (
   <AddLedgerAccountProvider onSuccess={onSuccess}>
     <Routes>
       <Route index element={<AddLedgerSelectNetwork />} />
