+++ apps/extension/src/ui/domains/Account/AccountAdd/types.ts
@@ -1,0 +1,1 @@
+export type AccountAddPageProps = { onSuccess: (address: string) => void }
